+++ lib/req/application.ex
@@ -6,14 +6,21 @@ defmodule Req.Application do
   @impl true
   def start(_type, _args) do
     children = [
-      Supervisor.child_spec({Finch, name: Req.FinchHTTP1, pools: %{default: [protocol: :http1]}},
-        id: Req.FinchHTTP1
-      ),
-      Supervisor.child_spec({Finch, name: Req.FinchHTTP2, pools: %{default: [protocol: :http2]}},
-        id: Req.FinchHTTP2
-      )
+      {DynamicSupervisor, name: Req.FinchSupervisor}
     ]
 
-    Supervisor.start_link(children, strategy: :one_for_one)
+    with {:ok, _sup} <- Supervisor.start_link(children, strategy: :one_for_one) do
+      {:ok, _} =
+        DynamicSupervisor.start_child(
+          Req.FinchSupervisor,
+          {Finch, name: Req.FinchSupervisor.HTTP1, pools: %{default: [protocol: :http1]}}
+        )
+
+      {:ok, _} =
+        DynamicSupervisor.start_child(
+          Req.FinchSupervisor,
+          {Finch, name: Req.FinchSupervisor.HTTP2, pools: %{default: [protocol: :http2]}}
+        )
+    end
   end
 end